Reorganize files into one cwt module
