+++ src/lib.rs
@@ -6,7 +6,6 @@
 /// COSE algorithms.
 #[allow(non_camel_case_types)]
 pub mod algorithm;
-pub mod claim;
 pub mod cwt;
 mod header_map;
 mod protected;
